pyproject.toml
[tool.poetry]
name = "grim_opt"
version = "0.2.2"
description = "Greenfield Renewable energy source Investment Model"
authors = ["Ni Wang <N.Wang@tudelft.nl>", "João Paulo Pizani Flor <J.P.PizaniFlor@tudelft.nl>"]
license = "MIT"

homepage = "https://gitlab.tudelft.nl/nwang1/grim"
repository = "https://gitlab.tudelft.nl/nwang1/grim"
readme = "README.md"


[tool.poetry.scripts]
grim_opt = "grim_opt.cli_app:main"

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
numpy = "^1.21.5"
pandas = "^1.2.4"
Pyomo = "^5.7.3"
openpyxl = "^3.0.7"
networkx = "^2.5.1"
geojson = "^2.5.0"
netCDF4 = "^1.5.6"
matplotlib = "^3.4.1"
Shapely = "^1.7.1"
seaborn = "^0.11.1"
scipy = "^1.6.2"
pyproj = "^3.0.1"
pyshp = "^2.1.3"
OWSLib = "^0.25.0"
Pillow = "^8.4.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
